Big simplification pass
Dropping reward strategies, it's unlikely I'll work on this project enough to need the structure for them any time soon.
Start of UI rework to make everything cleaner and clearer.
The text client is also being dropped, as it's currently just a debug tool, and could be handled via logging instead. Maybe I'll add it back in one day, but it's not really a goal of this project.